Added ssh and sudo client side wrappers
This resolves a few bugs in rsnapshot- namely a bug where "sudo rsync"
is considered a single file.

This also provides some enhanced security by locking down the sudo
rsync command to only work on the rsync_sender script, which prevents
anyone with access to the backup account from overwriting files on the
system.

All of this also ties into the SSH key by locking it down to a specific
command and host.