Separate sudo support into two parts
This change separates sudo support into two parts.  The first part is
configuring the suoders configuration with the saz/sudo module, and the
second part is configuring rsnapshot and the wrapper scripts to use
sudo.  This allows people that don't want to pull in a dependency on
saz/sudo to manually configure sudo and still have this module configure
rsnapshot for sudo.