Specify SSH key target to fix idempotency issue
If the target isn't specified, then sshkeys::set_authorized_key depends
on the homedir fact.  Since the user doesn't exist on the first run,
this fact won't exist and the first puppet run will fail.  Since we have
the home directory hard coded in the user creation anyway, go ahead and
provide that target to the sshkeys class to avoid the idempotency issue.