Update user.pp to allow IPv6
This change adds IPv6 addresses to the authorized_keys file of the backshots user. This is required, as OpenSSH 6.8 and up have changed the use of the configuration parameter "UseDNS" from Yes to No, which means that SSHd will no longer be doing DNS lookups.

Without this patch, a backup server that tries to connect using IPv6 to a server that has UseDNS set to 'no', the connection will fail.